pyproject.toml
[tool.poetry]
name = "numpy-nn-model"
version = "0.1.0"
description = ""
authors = ["AkiRusProd <akirusprod@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "1.24.0"
matplotlib = "3.5.1"
pillow = "9.3.0"
tqdm = "^4.66.1"
cupy-cuda12x = "^13.6.0"
requests = "^2.32.3"
tokenizers = "^0.19.1"
datasets = "^2.20.0"
ipywidgets = "^8.1.3"
ipympl = "^0.9.4"
nvtx = "^0.2.14"
pytest = "^9.0.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.7"
pre-commit = "^3.7.1"
mypy = "^1.10.0"
types-requests = "^2.32.0.20240602"
ipykernel = "^6.29.5"
torch = "^2.3.1"

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["venv", "tests"]

[tool.ruff.lint]
ignore = ["RUF002", "RUF003", "T201", "ERA001", "FBT002", "FBT001", "A001", "C901", "S403", "S301"]
fixable = ["I", "B", "RUF", "F"]
select = [
    "I",     # Isort rules
    "C90",   # Maccabe complexity
    "ASYNC", # flake8-async
    "S",     #  flake8-bandit
    "BLE",   # flake8-blind-except
    "B",     # flake8-bugbear
    "A",     # flake8-builtins
    "C4",    # flake8-comprehensions
    "T20",   # flake8-print
    # "RET",   # flake8-return
    "SLF", # flake8-self
    "RUF", # ruff
    "ERA", # eradicate
    "PTH", # flake8-use-pathlib
    "FBT", # flake8-boolean-trap
    "F",   # pyflakes
]
preview = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F403"]

[tool.mypy]
exclude = ["venv", "tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
